requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.3
django-admin-interface==0.19.2
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-discover-runner==1.0
django-easy-select2==1.5.8
django-extensions==3.2.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-localflavor==3.1
django-role-permissions==3.1.1
django-taggit==3.0.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
gunicorn==20.1.0
idna==3.3
importlib-resources==5.9.0
ipdb==0.13.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.10.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lxml==4.9.1
MarkupSafe==2.1.1
martor==1.6.16
matplotlib-inline==0.1.5
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-stdnum==1.17
pytz==2022.1
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.4.0
static3==0.7.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
traitlets==5.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.1
zipp==3.8.1
django-widget-tweaks==1.4.12